Modularize seed.js into a Folder.
